# packages/contracts/internal/core-api.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/chat/conversations/{id}/messages/{messageId}": {
      "put": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "editMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "deleteMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/messages/{messageId}/reaction": {
      "put": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "react",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatReactionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "removeReaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatReactionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/team-conversations": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "createTeamConversation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/sync-cursor": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "syncCursor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCursorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/chat/conversations": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "listConversations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatConversationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "createConversation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatConversationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/chat/conversations/{id}/route-plan": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "sendRoutePlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/route-plan/troca": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "trocaMotorista",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/route-plan/solicitar-troca": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "solicitarTroca",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/route-plan/solicitar-cancelamento": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "solicitarCancelamento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/route-plan/cancel": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "cancelRoutePlan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRoutePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/read": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "markAsRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/messages": {
      "get": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "listMessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "sendMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/conversations/{id}/messages/{messageId}/forward": {
      "post": {
        "tags": [
          "chat-controller"
        ],
        "operationId": "forwardMessage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForwardMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditMessageRequest": {
        "required": [
          "body"
        ],
        "type": "object",
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "senderUserId": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "lidoEm": {
            "type": "string",
            "format": "date-time"
          },
          "messageType": {
            "type": "string",
            "enum": [
              "TEXTO",
              "ATRIBUICAO_ROTA",
              "CANCELAMENTO_ROTA",
              "TROCA_MOTORISTA",
              "SOLICITACAO_CANCELAMENTO",
              "SOLICITACAO_TROCA_MOTORISTA",
              "SISTEMA"
            ]
          },
          "routePlanId": {
            "type": "string",
            "format": "uuid"
          },
          "stillOnServer": {
            "type": "boolean"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time"
          },
          "replyToMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "replyToBody": {
            "type": "string"
          },
          "replyToSenderUserId": {
            "type": "string",
            "format": "uuid"
          },
          "forwardedFromMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatReactionResponse"
            }
          }
        }
      },
      "ChatReactionResponse": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ReactRequest": {
        "required": [
          "emoji"
        ],
        "type": "object",
        "properties": {
          "emoji": {
            "maxLength": 8,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
      "CreateTeamConversationRequest": {
        "required": [
          "otherUserId"
        ],
        "type": "object",
        "properties": {
          "otherUserId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ChatConversationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "driverName": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "vehiclePlate": {
            "type": "string"
          },
          "otherParticipantUserId": {
            "type": "string",
            "format": "uuid"
          },
          "otherParticipantEmail": {
            "type": "string"
          },
          "otherParticipantRole": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastMessageBody": {
            "type": "string"
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SyncCursorRequest": {
        "required": [
          "deviceId",
          "syncedAt"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "syncedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateConversationRequest": {
        "required": [
          "driverId"
        ],
        "type": "object",
        "properties": {
          "driverId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SendRoutePlanRequest": {
        "required": [
          "routePlanId"
        ],
        "type": "object",
        "properties": {
          "routePlanId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SendMessageRequest": {
        "required": [
          "body"
        ],
        "type": "object",
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "replyToMessageId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ForwardMessageRequest": {
        "required": [
          "targetConversationId"
        ],
        "type": "object",
        "properties": {
          "targetConversationId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  }
}